Final finish all todo and create a README

+++ lib/domain/usecases/group/delete_group_usecase.dart
@@ -1,10 +1,12 @@
 import '../../repositories/i_group_repository.dart';
+import '../../repositories/i_enrollment_repository.dart';
 
 /// Use case for deleting a group
 class DeleteGroupUseCase {
   final IGroupRepository _groupRepository;
+  final IEnrollmentRepository _enrollmentRepository;
 
-  DeleteGroupUseCase(this._groupRepository);
+  DeleteGroupUseCase(this._groupRepository, this._enrollmentRepository);
 
   /// Execute the delete group use case
   Future<void> execute(String groupId) async {
@@ -12,8 +14,17 @@ class DeleteGroupUseCase {
       throw Exception('Group ID cannot be empty');
     }
 
-    // TODO: Check if group has enrollments before deleting
-    // If it does, prevent deletion or handle cascade delete
+    // Check if group has enrollments before deleting
+    final enrollments = await _enrollmentRepository.getEnrollmentsByGroup(
+      groupId,
+    );
+
+    if (enrollments.isNotEmpty) {
+      throw Exception(
+        'Cannot delete group: Group has ${enrollments.length} student(s) enrolled. '
+        'Please remove all students from the group before deleting.',
+      );
+    }
 
     return await _groupRepository.deleteGroup(groupId);
   }
