+++ lib/domain/usecases/semester/delete_semester_usecase.dart
@@ -1,18 +1,27 @@
 import '../../repositories/i_semester_repository.dart';
+import '../../repositories/i_course_repository.dart';
 
 /// Use case for deleting a semester
 class DeleteSemesterUseCase {
   final ISemesterRepository _repository;
+  final ICourseRepository _courseRepository;
 
-  DeleteSemesterUseCase(this._repository);
+  DeleteSemesterUseCase(this._repository, this._courseRepository);
 
   Future<void> execute(String semesterId) async {
     if (semesterId.isEmpty) {
       throw Exception('Semester ID cannot be empty');
     }
 
-    // TODO: Add check if semester has associated courses
-    // If it does, prevent deletion or cascade delete
+    // Check if semester has associated courses
+    final courses = await _courseRepository.getCoursesBySemester(semesterId);
+
+    if (courses.isNotEmpty) {
+      throw Exception(
+        'Cannot delete semester: Semester has ${courses.length} course(s) associated. '
+        'Please delete or reassign all courses before deleting the semester.',
+      );
+    }
 
     await _repository.deleteSemester(semesterId);
   }
